+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.3.0:
+ - Added support for custom word lists (`-words).
+ - `-theme` now accepts a path.
+ - Added `~/.tt/themes` and `~/.tt/words`.
+ - Scrapped ~/.ttrc in favour of aliases/flags.
+ - Included more default word lists. (`-list words`)
+
 # 0.2.2:
  - Modified -g to correspond to the number of groups rather than the group size.
  - Added -multi